+++ README.md
@@ -33,14 +33,17 @@ MOD_CHAT_ID=-1002527730048
 - `ADMIN_ID` – идентификатор администратора
 - `MOD_CHAT_ID` – ID чата модерации
 
-## Запуск
-
 После настройки переменных окружения запустите бота командой:
 ```bash
 python bot.py
 ```
 
 Бот пересылает все личные сообщения в чат модерации и отвечает пользователю об успешной отправке.
+
+## Структура проекта
+
+Исходный код бота находится в каталоге `app`.
+Основная точка входа — файл `bot.py`, который подключает обработчики из `app.handlers`.
 
 ## Предложение контента
